Adding changes based on code review
